# openapi.yaml
openapi: 3.1.0
info:
  title: Astradio API
  version: "1.0"
servers:
  - url: https://api.astradio.io/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Track:
      type: object
      required: [id, owner_id, mode, genre, duration_sec]
      properties:
        id: { type: string, format: ulid }
        owner_id: { type: string, format: ulid }
        mode: { type: string, enum: [clusters, elemental, lunar] }
        genre: { type: string }
        source: { type: string, enum: [natal, transit, overlay] }
        duration_sec: { type: integer }
        key_signature: { type: string }
        bpm: { type: integer }
        waveform_url: { type: string }
        preview_url: { type: string }
        og_image_url: { type: string }
        timeline_json: { type: object }
security:
  - bearerAuth: []
paths:
...
  /tracks/render:
    post:
      summary: Enqueue a render
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chartData, mode, genre]
              properties:
                chartData: { type: object }
                mode: { type: string, enum: [clusters, elemental, lunar] }
                genre: { type: string }
      responses:
        "202": { description: Enqueued }
        "200": { description: Deduped existing, returns Track, content: { application/json: { schema: { $ref: "#/components/schemas/Track" } } } }
  /tracks/{id}:
    get:
      summary: Get track by id
      parameters: [{ name: id, in: path, required: true, schema: {type: string} }]
      responses:
        "200":
          description: Track
          content: { application/json: { schema: { $ref: "#/components/schemas/Track" } } }
...